feat: prepare package ffg
