+++ .gitignore
@@ -4,3 +4,6 @@ __pycache__/
 *.dot
 out/
 *.env
+build/
+ffg.egg-info
+dist/
