+++ ffg/__init__.py
@@ -19,32 +19,3 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
-
-
-import os
-from typing import List
-
-from src.operators.generic import Operator
-from src.visitors.dot_visitor import DotVisitor
-from src.visitors.rewrite_visitor import RewriteVisitor
-
-
-def emit(trees: List[Operator], output_dir):
-    """
-    Emits multiple fusion functions and its inverses to the dot format.
-    Each function is placed in a separate subdirectory inside ./out.
-    """
-    dot_printer = DotVisitor()
-    rewriter = RewriteVisitor()
-    for i, tree in enumerate(trees):
-        directory = os.path.join(output_dir, f"dot_{i}")
-        os.makedirs(directory, exist_ok=True)
-
-        with open(os.path.join(directory, "z.dot"), 'w', encoding='utf-8') as file:
-            digraph = tree.accept(dot_printer)
-            print(digraph, file=file)
-
-        for inv_i, inverse in enumerate(tree.accept(rewriter)):
-            with open(os.path.join(directory, f"inv_{inv_i}.dot"), 'w', encoding='utf-8') as file:
-                digraph = inverse.accept(dot_printer)
-                print(digraph, file=file)
