+++ ffg/emitter/__init__.py
