+++ ffg/operators/__init__.py
