+++ ffg/operators/generic.py
