+++ ffg/operators/real_theory.py
@@ -25,7 +25,7 @@
 # Look at the README to learn more.
 
 from abc import ABC, abstractmethod
-from src.operators.generic import IntegerOperator, RealOperator
+from ffg.operators.generic import IntegerOperator, RealOperator
 
 
 class RealNegation(RealOperator):
