+++ ffg/visitors/printer_visitor.py
@@ -21,91 +21,91 @@
 # SOFTWARE.
 
 
-from src.operators.boolean_theory import *
-from src.operators.integer_theory import *
-from src.operators.real_theory import *
-from src.operators.string_theory import *
-from src.operators.bitvector_theory import *
+from ffg.operators.boolean_theory import *
+from ffg.operators.integer_theory import *
+from ffg.operators.real_theory import *
+from ffg.operators.string_theory import *
+from ffg.operators.bitvector_theory import *
 
 
-class VariableVisitor(BooleanVisitor, IntegerVisitor, RealVisitor, StringVisitor, BitVectorVisitor):
-    def visit_boolean_not(self, operator: BooleanNot):
-        return self._visit_operator(operator, 1)
-
+class PrinterVisitor(BooleanVisitor, IntegerVisitor, RealVisitor, StringVisitor, BitVectorVisitor):
     def visit_boolean_xor(self, operator: BooleanXor):
-        return self._visit_operator(operator, 2)
+        return f"(xor {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
-    def visit_boolean_variable(self, operator: BooleanVariable):
-        return {operator.name: "Bool"}
+    def visit_boolean_not(self, operator: BooleanNot):
+        return f"(not {operator.operator_1.accept(self)})"
 
     def visit_boolean_constant(self, operator: BooleanConstant):
-        return {}
+        return operator.name
+
+    def visit_boolean_variable(self, operator: BooleanVariable):
+        return operator.name
 
     def visit_boolean_literal(self, operator: BooleanLiteral):
-        return {}
+        return str(operator.value).lower()
 
     def visit_boolean_equality(self, operator: BooleanEquality):
-        return self._visit_operator(operator, 2)
+        return f"(= {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def visit_integer_negation(self, operator: IntegerNegation):
-        return self._visit_operator(operator, 1)
+        return f"(- {operator.operator_1.accept(self)})"
 
     def visit_integer_addition(self, operator: IntegerAddition):
-        return self._visit_operator(operator, 2)
+        return f"(+ {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def visit_integer_subtraction(self, operator: IntegerSubtraction):
-        return self._visit_operator(operator, 2)
+        return f"(- {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def visit_integer_multiplication(self, operator: IntegerMultiplication):
-        return self._visit_operator(operator, 2)
+        return f"(* {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def visit_integer_division(self, operator: IntegerDivision):
-        return self._visit_operator(operator, 2)
-
-    def visit_integer_variable(self, operator: IntegerVariable):
-        return {operator.name: "Int"}
+        return f"(div {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def visit_integer_constant(self, operator: IntegerConstant):
-        return {}
+        return operator.name
+
+    def visit_integer_variable(self, operator: IntegerVariable):
+        return operator.name
 
     def visit_integer_literal(self, operator: IntegerLiteral):
-        return {}
+        return str(operator.value)
 
     def visit_integer_equality(self, operator: IntegerEquality):
-        return self._visit_operator(operator, 2)
+        return f"(= {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def visit_real_negation(self, operator: RealNegation):
-        return self._visit_operator(operator, 1)
+        return f"(- {operator.operator_1.accept(self)})"
 
     def visit_real_addition(self, operator: RealAddition):
-        return self._visit_operator(operator, 2)
+        return f"(+ {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def visit_real_subtraction(self, operator: RealSubtraction):
-        return self._visit_operator(operator, 2)
+        return f"(- {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def visit_real_multiplication(self, operator: RealMultiplication):
-        return self._visit_operator(operator, 2)
+        return f"(* {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def visit_real_division(self, operator: RealDivision):
-        return self._visit_operator(operator, 2)
+        return f"(/ {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def visit_integer_to_real(self, operator: IntegerToReal):
-        return self._visit_operator(operator, 1)
-
-    def visit_real_variable(self, operator: RealVariable):
-        return {operator.name: "Real"}
+        return f"(to_real {operator.operator_1.accept(self)})"
 
     def visit_real_constant(self, operator: RealConstant):
-        return {}
+        return operator.name
+
+    def visit_real_variable(self, operator: RealVariable):
+        return operator.name
 
     def visit_real_literal(self, operator: RealLiteral):
-        return {}
+        return str(operator.value)
 
     def visit_real_equality(self, operator: RealEquality):
-        return self._visit_operator(operator, 2)
+        return f"(= {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def _visit_string_concatenation(self, operator: StringOperator):
-        return self._visit_operator(operator, 2)
+        return f"(str.++ {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def visit_string_concatenation1n1(self, operator: StringConcatenation1n1):
         return self._visit_string_concatenation(operator)
@@ -126,62 +126,59 @@ class VariableVisitor(BooleanVisitor, IntegerVisitor, RealVisitor, StringVisitor
         return self._visit_string_concatenation(operator)
 
     def visit_string_length(self, operator: StringLength):
-        return operator.operator_1.accept(self)
+        return f"(str.len {operator.operator_1.accept(self)})"
 
     def visit_string_indexof(self, operator: StringIndexof):
-        return self._visit_operator(operator, 3)
+        return f"(str.indexof {operator.operator_1.accept(self)} " \
+               f"{operator.operator_2.accept(self)} {operator.operator_3.accept(self)})"
 
     def visit_real_to_integer(self, operator: RealToInteger):
-        return self._visit_operator(operator, 1)
+        return f"(to_int {operator.operator_1.accept(self)})"
 
     def visit_substring(self, operator: Substring):
-        return self._visit_operator(operator, 3)
+        return f"(str.substr {operator.operator_1.accept(self)} " \
+               f"{operator.operator_2.accept(self)} {operator.operator_3.accept(self)})"
 
     def visit_string_replacement(self, operator: StringReplacement):
-        return self._visit_operator(operator, 3)
+        return f"(str.replace {operator.operator_1.accept(self)} " \
+               f"{operator.operator_2.accept(self)} {operator.operator_3.accept(self)})"
 
     def visit_string_variable(self, operator: StringVariable):
-        return {operator.name: "String"}
+        return operator.name
 
     def visit_string_constant(self, operator: StringLiteral):
-        return {}
+        return operator.name
 
     def visit_string_literal(self, operator: StringLiteral):
-        return {}
+        return f"\"{operator.value}\""
 
     def visit_string_equality(self, operator: StringEquality):
-        return self._visit_operator(operator, 2)
+        return f"(= {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def visit_bit_vector_not(self, operator: BitVectorNot):
-        return self._visit_operator(operator, 1)
+        return f"(bvnot {operator.operator_1.accept(self)})"
 
     def visit_bit_vector_negation(self, operator: BitVectorNegation):
-        return self._visit_operator(operator, 1)
+        return f"(bvneg {operator.operator_1.accept(self)})"
 
     def visit_bit_vector_xor(self, operator: BitVectorXor):
-        return self._visit_operator(operator, 2)
+        return f"(bvxor {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def visit_bit_vector_concatenation(self, operator: BitVectorConcatenation):
-        return self._visit_operator(operator, 2)
+        return f"(concat {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
 
     def visit_bit_vector_extraction(self, operator: BitVectorExtraction):
-        return self._visit_operator(operator, 3)
+        return f"((_ extract {operator.operator_3.accept(self)} {operator.operator_2.accept(self)}) " \
+               f"{operator.operator_1.accept(self)})"
 
     def visit_bit_vector_variable(self, operator: BitVectorVariable):
-        return {operator.name: f"(_ BitVec {operator.size})"}
+        return operator.name
 
     def visit_bit_vector_constant(self, operator: BitVectorConstant):
-        return {}
+        return operator.name
 
     def visit_bit_vector_literal(self, operator: BitVectorLiteral):
-        return {}
+        return f"#b{''.join([str(bit) for bit in operator.value])}"
 
     def visit_bit_vector_equality(self, operator: BitVectorEquality):
-        return self._visit_operator(operator, 2)
-
-    def _visit_operator(self, operator, arity):
-        out = dict()
-        for i in range(arity):
-            op_out = getattr(operator, f'operator_{i + 1}').accept(self)
-            out.update(op_out)
-        return out
+        return f"(= {operator.operator_1.accept(self)} {operator.operator_2.accept(self)})"
