+++ setup.py
@@ -19,3 +19,14 @@
 # LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 # SOFTWARE.
+
+from setuptools import find_packages, setup
+
+setup(
+    name='ffg',
+    packages=find_packages(exclude=['tests']),
+    version='0.1.0',
+    description='A generator of fusion functions for Semantic Fusion (YinYang).',
+    author='Nicola Dardanis, Lucas Weitzendorf',
+    license='MIT',
+)
